Image_generator/requirements.txt
streamlit
python-dotenv
requests
pillow
